added code for binning behaviors
